Preparation for saving wildignore per buffer
autoload/gitignore/git.vim: Some git functions starting with gitignore#git#root() to find the git root of the current file ('%').  Git roots are cached in a buffer local variable, b:gitignore_root, and a map of absolute directory name to absolute git root directory is also maintained to make finding the git root for a new buffer as efficient as possible.